🥂 To New Beginnings!! 🥳
A lot happened before this, but I decided to erase the whole git history in preparations to make this repository public. There were a lot of API keys hardcoded in the code, that needed to be removed (even tough they are useless if anyone is reading this).

I have been working on this since May 25, 2025.